providers apod: reorg and tests (more)
I found out that the old pages are giving errors due an old layout without center tag. But they also have very small images (due their age), so it doesn't work for wallpaper ones. Examined others errors and find out that all we are interested are in the pages we can parse, so I'll just go ignore all the ones with errors and carry on. Also will study  downloading the images to check about when the sizes begun to be ok.